refactor: move server logic to src
